pref: move set_informational_units and git_prefs_units_set to pref.cpp
Both are preference-layer concerns that sit alongside git_prefs and
set_git_prefs. They ended up in dive.cpp only because
set_informational_units() was already there. Move them to their natural
home so that all git_prefs-related state and logic lives in one place.

No functional change; core/pref.h, core/load-git.cpp,
mobile-widgets/qmlmanager.cpp, and tests/testqPrefUnits.cpp are
unchanged.
